refactor: extract deliver_webhook into shared webhook module
