+++ src/monitor.rs
@@ -2,7 +2,7 @@ use crate::graph::GraphStore;
 use crate::state::StateDb;
 use serde::{Deserialize, Serialize};
 use std::sync::Arc;
-use std::time::Duration;
+
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum WatcherAction {
@@ -103,7 +103,7 @@ impl Monitor {
                             "timestamp": chrono::Utc::now().to_rfc3339(),
                         });
                         tokio::spawn(async move {
-                            let _ = deliver_webhook(&url, headers.as_deref(), &payload).await;
+                            let _ = crate::webhook::deliver_webhook(&url, headers.as_deref(), &payload).await;
                         });
                     }
                 }
@@ -218,29 +218,4 @@ impl Monitor {
             Err(_) => 0.0,
         }
     }
-}
-
-/// Fire-and-forget webhook delivery with 10s timeout.
-async fn deliver_webhook(
-    url: &str,
-    headers_json: Option<&str>,
-    payload: &serde_json::Value,
-) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    let client = reqwest::Client::builder()
-        .timeout(Duration::from_secs(10))
-        .build()?;
-    let mut req = client.post(url).json(payload);
-    if let Some(hdr_json) = headers_json {
-        if let Ok(map) = serde_json::from_str::<std::collections::HashMap<String, String>>(hdr_json) {
-            for (k, v) in map {
-                req = req.header(&k, &v);
-            }
-        }
-    }
-    let resp = req.send().await?;
-    let status = resp.status();
-    if !status.is_success() {
-        eprintln!("Webhook to {} returned {}", url, status);
-    }
-    Ok(())
 }
