+++ src/webhook.rs
@@ -1,0 +1,26 @@
+use std::time::Duration;
+
+/// Fire-and-forget webhook delivery with 10s timeout.
+pub async fn deliver_webhook(
+    url: &str,
+    headers_json: Option<&str>,
+    payload: &serde_json::Value,
+) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+    let client = reqwest::Client::builder()
+        .timeout(Duration::from_secs(10))
+        .build()?;
+    let mut req = client.post(url).json(payload);
+    if let Some(hdr_json) = headers_json {
+        if let Ok(map) = serde_json::from_str::<std::collections::HashMap<String, String>>(hdr_json) {
+            for (k, v) in map {
+                req = req.header(&k, &v);
+            }
+        }
+    }
+    let resp = req.send().await?;
+    let status = resp.status();
+    if !status.is_success() {
+        eprintln!("Webhook to {} returned {}", url, status);
+    }
+    Ok(())
+}
